fix kernel NULL pointer dereference in diff_buffer_free()
The bug was confirmed in the case when we were successfully able to
allocate memory for the struct diff_buffer, but could not allocate
pages for it in the proper amount. When diff_buffer_free() is called,
pages are freed, the number of which is determined by the nr_pages.

In the fix, the nr_pages defines the number of successfully allocated
pages.