use free_diff_buffer_sem instead free_diff_buffers_count
The logic of buffer allocation has been changed.
Previously, the freq_diff_buffer_pool_size parameter determined the
number of available buffers in the cache, but did not limit their total
number. When there were insufficient buffers in the cache, an allocation
was performed. This could lead to excessive memory consumption, lack of
free pages, and a crash of the kernel.
Now, the freq_diff_buffer_pool_size parameter limits the total number of
available buffers per block device.